Refactor download handling to hook
