+++ App.js
@@ -10,30 +10,26 @@
 import 'react-native-url-polyfill/auto';
 
 import { Ionicons } from '@expo/vector-icons';
-import { getPlaystateApi } from '@jellyfin/sdk/lib/utils/api/playstate-api';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { NavigationContainer } from '@react-navigation/native';
 import { Asset } from 'expo-asset';
-import * as FileSystem from 'expo-file-system';
 import * as Font from 'expo-font';
 import * as ScreenOrientation from 'expo-screen-orientation';
 import * as SplashScreen from 'expo-splash-screen';
 import { StatusBar } from 'expo-status-bar';
 import PropTypes from 'prop-types';
 import React, { useContext, useEffect, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { Alert, useColorScheme } from 'react-native';
+import { useColorScheme } from 'react-native';
 import { ThemeContext, ThemeProvider } from 'react-native-elements';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 
 import ThemeSwitcher from './components/ThemeSwitcher';
-import { DownloadStatus } from './constants/DownloadStatus';
+import { useDownloadHandler } from './hooks/useDownloadHandler';
 import { useIsHydrated } from './hooks/useHydrated';
 import { useStores } from './hooks/useStores';
 import { fromStorageObject } from './models/DownloadModel';
 import ServerModel from './models/ServerModel';
 import RootNavigator from './navigation/RootNavigator';
-import { ensurePathExists } from './utils/File';
 import StaticScriptLoader from './utils/StaticScriptLoader';
 
 // Import i18n configuration
@@ -52,7 +48,9 @@ const App = ({ skipLoadingScreen }) => {
 	const { theme } = useContext(ThemeContext);
 	const isHydrated = useIsHydrated();
 	const colorScheme = useColorScheme();
-	const { t } = useTranslation();
+
+	// Initialize download hook
+	useDownloadHandler();
 
 	// Store the system color scheme for automatic theme switching
 	useEffect(() => {
@@ -192,73 +190,6 @@ const App = ({ skipLoadingScreen }) => {
 		// Update the screen orientation
 		updateScreenOrientation();
 	}, [ rootStore.isFullscreen ]);
-
-	useEffect(() => {
-		const downloadFile = async (download) => {
-			console.debug('[App] downloading "%s"', download.filename);
-			// For transcoded downloads we force .mp4
-			if (!download.extension) {
-				download.extension = '.mp4';
-				downloadStore.update(download);
-			}
-			await ensurePathExists(download.localPath);
-
-			const url = download.getStreamUrl(rootStore.deviceId);
-
-			const resumable = FileSystem.createDownloadResumable(
-				url.toString(),
-				download.uri,
-				{},
-				(/*{ totalBytesWritten }*/) => {
-					// FIXME: We should save the download progress in the model for display
-					// but this needs throttling
-				}
-			);
-
-			// TODO: The resumable should be saved to allow pausing/resuming downloads
-
-			// Download the file
-			try {
-				download.status = DownloadStatus.Downloading;
-				downloadStore.update(download);
-				await resumable.downloadAsync();
-				download.status = DownloadStatus.Complete;
-			} catch (e) {
-				console.error('[App] Download failed', e);
-				Alert.alert(
-					t('alerts.downloadFailed.title'),
-					t('alerts.downloadFailed.description', { title: download.title })
-				);
-
-				// TODO: If a download fails, we should probably remove it from the queue
-				download.status = DownloadStatus.Failed;
-			}
-
-			// Push the state update to the store
-			downloadStore.update(download);
-
-			// Report download has stopped
-			const serverUrl = download.serverUrl.endsWith('/') ? download.serverUrl.slice(0, -1) : download.serverUrl;
-			const api = rootStore.getSdk().createApi(serverUrl, download.apiKey);
-			console.log('[App] Reporting download stopped', download.sessionId);
-			getPlaystateApi(api)
-				.reportPlaybackStopped({
-					playbackStopInfo: {
-						PlaySessionId: download.sessionId
-					}
-				})
-				.catch(err => {
-					console.error('[App] Failed reporting download stopped', err.response || err.request || err.message);
-				});
-		};
-
-		downloadStore.downloads
-			.forEach(download => {
-				if (download.status === DownloadStatus.Pending) {
-					downloadFile(download);
-				}
-			});
-	}, [ rootStore.deviceId, downloadStore.downloads.size ]);
 
 	if (!(isSplashReady && isStoresReady) && !skipLoadingScreen) {
 		return null;
